fix namings, move tests, change float->int return type
